+++ src/app/app.component.ts
@@ -1,5 +1,5 @@
 import { Component } from '@angular/core';
-import { DataService } from './data.service';
+import { DataService } from './services/data/data.service';
 
 @Component({
   selector: 'app-root',
