+++ src/app/components/table/table.component.html
@@ -2,17 +2,17 @@
   <thead>
     <tr>
       <th>
-        <button mat-raised-button (click)="reloadButton()" class="botonCargar">
+        <button mat-raised-button (click)="getTenCountries()" class="botonCargar">
           Cargar nuevos Paises
         </button>
       </th>
       <th>
-        <button mat-raised-button (click)="sortByCountry()">
+        <button mat-raised-button (click)="sortByCountryButton()">
           Ordenar por Pais ↓
         </button>
       </th>
       <th>
-        <button mat-raised-button (click)="sortByCity()">
+        <button mat-raised-button (click)="sortByCityButton()">
           Ordenar por Capital ↓
         </button>
       </th>
